GnuPG-mini-HOWTO.md: (Re)write first chapter
Okay, I had a crash and associated filesystem corruption.  Sadly, the
backup wasn't quite current so I lost some data not yet pushed.  So I
have to rewrite the GnuPG mini-HOWTO all over again.  The hardest part
of this is the intro chapter since from there it's mostly a matter of
teaching the basic commands, key maintenance, and a discussion about the
web of trust.

Honestly, this first chapter is actually better than the one I wrote
last time.  ;)