GnuPG-mini-HOWTO: begin rewriting
Seems that the explanations are too long to follow, even if the concepts
are simple enough.  Let's break it down a little more for the sake of
brevity.  If this needs to be expanded upon, I'm sure someone will let
me know where more info is needed.

This leaves the second part in a bit inconsistent state since I've
folded part of it into the first part.  Next commit.  Bedtime for now.